+++ src/Order/Changelog.md
@@ -2,6 +2,11 @@
 
 ## 3.x Series
 
+## Unreleased
+##### 2023-XX-YY
+
+- Added the `ofUser()` scope to the base Order model
+
 ## 3.7.0
 ##### 2023-04-04